# Optimize completion query, benchmarks, JobInsertFullMany