--- a/riverdriver/riverdatabasesql/internal/dbsqlc/river_job.sql.go
+++ b/riverdriver/riverdatabasesql/internal/dbsqlc/river_job.sql.go
@@ -1192,80 +1192,81 @@ func (q *Queries) JobSchedule(ctx context.Context, db DBTX, arg *JobSchedulePara
 const jobSetStateIfRunningMany = `-- name: JobSetStateIfRunningMany :many
 WITH job_input AS (
     SELECT
-        unnest($1::bigint[]) AS id,
-        unnest($2::boolean[]) AS attempt_do_update,
-        unnest($3::int[]) AS attempt,
-        unnest($4::boolean[]) AS errors_do_update,
-        unnest($5::jsonb[]) AS errors,
+        unnest($1::bigint[])                     AS id,
+        unnest($2::boolean[])      AS attempt_do_update,
+        unnest($3::int[])                    AS attempt,
+        unnest($4::boolean[])       AS errors_do_update,
+        unnest($5::jsonb[])                   AS errors,
         unnest($6::boolean[]) AS finalized_at_do_update,
-        unnest($7::timestamptz[]) AS finalized_at,
-        unnest($8::boolean[]) AS metadata_do_merge,
-        unnest($9::jsonb[]) AS metadata_updates,
+        unnest($7::timestamptz[])       AS finalized_at,
+        unnest($8::boolean[])      AS metadata_do_merge,
+        unnest($9::jsonb[])         AS metadata_updates,
         unnest($10::boolean[]) AS scheduled_at_do_update,
-        unnest($11::timestamptz[]) AS scheduled_at,
+        unnest($11::timestamptz[])       AS scheduled_at,
         -- To avoid requiring pgx users to register the OID of the river_job_state[]
         -- type, we cast the array to text[] and then to river_job_state.
         unnest($12::text[])::/* TEMPLATE: schema */river_job_state AS state
 ),
-job_to_update AS (
-    SELECT
-        river_job.id,
-        job_input.attempt,
-        job_input.attempt_do_update,
-        job_input.errors,
-        job_input.errors_do_update,
-        job_input.finalized_at,
-        job_input.finalized_at_do_update,
-        job_input.metadata_do_merge,
-        job_input.metadata_updates,
-        job_input.scheduled_at,
-        job_input.scheduled_at_do_update,
-        (job_input.state IN ('retryable', 'scheduled') AND river_job.metadata ? 'cancel_attempted_at') AS should_cancel,
-        job_input.state
-    FROM /* TEMPLATE: schema */river_job
-    JOIN job_input ON river_job.id = job_input.id
-    WHERE river_job.state = 'running' OR job_input.metadata_do_merge
-    FOR UPDATE
-),
-updated_running AS (
+updated AS (
     UPDATE /* TEMPLATE: schema */river_job
     SET
-        attempt      = CASE WHEN NOT job_to_update.should_cancel AND job_to_update.attempt_do_update THEN job_to_update.attempt
-                            ELSE river_job.attempt END,
-        errors       = CASE WHEN job_to_update.errors_do_update THEN array_append(river_job.errors, job_to_update.errors)
-                            ELSE river_job.errors END,
-        finalized_at = CASE WHEN job_to_update.should_cancel THEN coalesce($13::timestamptz, now())
-                            WHEN job_to_update.finalized_at_do_update THEN job_to_update.finalized_at
-                            ELSE river_job.finalized_at END,
-        metadata     = CASE WHEN job_to_update.metadata_do_merge
-                            THEN river_job.metadata || job_to_update.metadata_updates
-                            ELSE river_job.metadata END,
-        scheduled_at = CASE WHEN NOT job_to_update.should_cancel AND job_to_update.scheduled_at_do_update THEN job_to_update.scheduled_at
-                            ELSE river_job.scheduled_at END,
-        state        = CASE WHEN job_to_update.should_cancel THEN 'cancelled'::/* TEMPLATE: schema */river_job_state
-                            ELSE job_to_update.state END
-    FROM job_to_update
-    WHERE river_job.id = job_to_update.id
-        AND river_job.state = 'running'
-    RETURNING river_job.id, river_job.args, river_job.attempt, river_job.attempted_at, river_job.attempted_by, river_job.created_at, river_job.errors, river_job.finalized_at, river_job.kind, river_job.max_attempts, river_job.metadata, river_job.priority, river_job.queue, river_job.state, river_job.scheduled_at, river_job.tags, river_job.unique_key, river_job.unique_states
-),
-updated_metadata_only AS (
-    UPDATE /* TEMPLATE: schema */river_job
-    SET metadata = river_job.metadata || job_to_update.metadata_updates
-    FROM job_to_update
-    WHERE river_job.id = job_to_update.id
-        AND river_job.id NOT IN (SELECT id FROM updated_running)
-        AND river_job.state != 'running'
-        AND job_to_update.metadata_do_merge
+        attempt = CASE
+            WHEN river_job.state = 'running'
+                 AND NOT (job_input.state IN ('retryable','scheduled') AND river_job.metadata ? 'cancel_attempted_at')
+                 AND job_input.attempt_do_update
+            THEN job_input.attempt
+            ELSE river_job.attempt
+        END,
+        errors = CASE
+            WHEN river_job.state = 'running'
+                 AND job_input.errors_do_update
+            THEN array_append(river_job.errors, job_input.errors)
+            ELSE river_job.errors
+        END,
+        finalized_at = CASE
+            WHEN river_job.state = 'running'
+                 AND (job_input.state IN ('retryable','scheduled') AND river_job.metadata ? 'cancel_attempted_at')
+            THEN coalesce($13::timestamptz, now())
+            WHEN river_job.state = 'running'
+                 AND job_input.finalized_at_do_update
+            THEN job_input.finalized_at
+            ELSE river_job.finalized_at
+        END,
+        metadata = CASE
+            WHEN job_input.metadata_do_merge
+            THEN river_job.metadata || job_input.metadata_updates
+            ELSE river_job.metadata
+        END,
+        scheduled_at = CASE
+            WHEN river_job.state = 'running'
+                 AND NOT (job_input.state IN ('retryable','scheduled') AND river_job.metadata ? 'cancel_attempted_at')
+                 AND job_input.scheduled_at_do_update
+            THEN job_input.scheduled_at
+            ELSE river_job.scheduled_at
+        END,
+        state = CASE
+            WHEN river_job.state = 'running'
+                 AND (job_input.state IN ('retryable','scheduled') AND river_job.metadata ? 'cancel_attempted_at')
+            THEN 'cancelled'::/* TEMPLATE: schema */river_job_state
+            WHEN river_job.state = 'running'
+            THEN job_input.state
+            ELSE river_job.state
+        END
+    FROM job_input
+    WHERE river_job.id = job_input.id
+      AND (river_job.state = 'running' OR job_input.metadata_do_merge)
     RETURNING river_job.id, river_job.args, river_job.attempt, river_job.attempted_at, river_job.attempted_by, river_job.created_at, river_job.errors, river_job.finalized_at, river_job.kind, river_job.max_attempts, river_job.metadata, river_job.priority, river_job.queue, river_job.state, river_job.scheduled_at, river_job.tags, river_job.unique_key, river_job.unique_states
 )
-SELECT id, args, attempt, attempted_at, attempted_by, created_at, errors, finalized_at, kind, max_attempts, metadata, priority, queue, state, scheduled_at, tags, unique_key, unique_states
+SELECT river_job.id, river_job.args, river_job.attempt, river_job.attempted_at, river_job.attempted_by, river_job.created_at, river_job.errors, river_job.finalized_at, river_job.kind, river_job.max_attempts, river_job.metadata, river_job.priority, river_job.queue, river_job.state, river_job.scheduled_at, river_job.tags, river_job.unique_key, river_job.unique_states
 FROM /* TEMPLATE: schema */river_job
-WHERE id IN (SELECT id FROM job_input)
-    AND id NOT IN (SELECT id FROM updated_metadata_only)
-    AND id NOT IN (SELECT id FROM updated_running)
-UNION SELECT id, args, attempt, attempted_at, attempted_by, created_at, errors, finalized_at, kind, max_attempts, metadata, priority, queue, state, scheduled_at, tags, unique_key, unique_states FROM updated_metadata_only
-UNION SELECT id, args, attempt, attempted_at, attempted_by, created_at, errors, finalized_at, kind, max_attempts, metadata, priority, queue, state, scheduled_at, tags, unique_key, unique_states FROM updated_running
+JOIN job_input ON river_job.id = job_input.id
+WHERE NOT EXISTS (
+    SELECT 1
+    FROM updated
+    WHERE updated.id = river_job.id
+)
+UNION ALL
+SELECT id, args, attempt, attempted_at, attempted_by, created_at, errors, finalized_at, kind, max_attempts, metadata, priority, queue, state, scheduled_at, tags, unique_key, unique_states FROM updated
 `
 
 type JobSetStateIfRunningManyParams struct {