As part of a recent experiment, I was paying a lot of attention to the job completion query. I did some analysis of it and figured out that this refactoring can speed it up by about 15%. It also appears to increase throughput significantly.

In the process of figuring this out I decided to start a driver benchmark suite similar to our test suites. I also needed access to `JobInsertFullMany` to insert many jobs in a variety of states, so I pulled that over from Pro.

## Query stats

Collected stats for benchmark runs, resetting between each. 

### master

```
       total_min        |        avg_ms         | calls |                                                                                                                                                                                                              query                                                  
                                                                                                                                                             
------------------------+-----------------------+-------+---------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
-------------------------------------------------------------------------------------------------------------------------------------------------------------
    0.24928899236666663 |    57.528229007692325 |   260 | -- name: JobSetStateIfRunningMany :many       
    0.23806803375000005 |    53.699556484962415 |   266 | -- name: JobSetStateIfRunningMany :many                                                                                                                                                                                                                                                                                                                                                                                         +
```

### With this change

```
       total_min        |         avg_ms         | calls |                                                                                                                                                                                                              query                                                                                                                                                                                                               
------------------------+------------------------+-------+----------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
    0.20998258465000008 |      46.31968779044117 |   272 | -- name: JobSetStateIfRunningMany :many                                                                                                                                                                                                                                                                                                                                                                                         +
     0.2078785631666666 |     47.066844490566034 |   265 | -- name: JobSetStateIfRunningMany :many                                                                                                                                                                                                                                                                                                                                                                                         +
```

## Benchmark

These benchmarks are completing 2000 jobs using a variety of job states to hit all clauses in the query.

### On master

```
$ go test -benchmem -run=^$ -bench ^BenchmarkDriverRiverPgxV5$ github.com/riverqueue/river -benchtime 30s
goos: darwin
goarch: arm64
pkg: github.com/riverqueue/river
cpu: Apple M4 Pro
BenchmarkDriverRiverPgxV5/JobSetStateIfRunningMany-14         	      33	  31044995 ns/op	 4605439 B/op	   53534 allocs/op
BenchmarkDriverRiverPgxV5/JobSetStateIfRunningMany-14         	      36	  34336087 ns/op	 4600520 B/op	   53532 allocs/op
BenchmarkDriverRiverPgxV5/JobSetStateIfRunningMany-14                987   34195653 ns/op         4596863 B/op      53533 allocs/op
```

### With this change

```
goos: darwin
goarch: arm64
pkg: github.com/riverqueue/river
cpu: Apple M4 Pro
BenchmarkDriverRiverPgxV5/JobSetStateIfRunningMany-14         	      36	  30853128 ns/op	 4595252 B/op	   53530 allocs/op
BenchmarkDriverRiverPgxV5/JobSetStateIfRunningMany-14               1131   32302842 ns/op         4595874 B/op      53532 allocs/op
```

## riverbench

The overall results are inconsistent as I run this several times (as always) but the pattern is consistent: this change bumps throughput by around 8-10k jobs/sec.

### On master

```
➜  river git:(bg-producer-completion) ✗ go run ./cmd/river bench --database-url "postgres://localhost/river_test" --num-total-jobs 1_000_000
bench: jobs worked [          0 ], inserted [    1000000 ], job/sec [        0.0 ] [0s]
May 17 14:27:43.144 WRN jobcompleter.BatchCompleter: Hit maximum backlog; completions will wait until below threshold max_backlog=20000
May 17 14:27:44.291 WRN jobcompleter.BatchCompleter: Hit maximum backlog; completions will wait until below threshold max_backlog=20000
bench: jobs worked [     106645 ], inserted [          0 ], job/sec [    53322.5 ] [2s]
May 17 14:27:45.349 WRN jobcompleter.BatchCompleter: Hit maximum backlog; completions will wait until below threshold max_backlog=20000
bench: jobs worked [     112089 ], inserted [          0 ], job/sec [    56044.5 ] [2s]
May 17 14:27:46.631 WRN jobcompleter.BatchCompleter: Hit maximum backlog; completions will wait until below threshold max_backlog=20000
May 17 14:27:47.895 WRN jobcompleter.BatchCompleter: Hit maximum backlog; completions will wait until below threshold max_backlog=20000
bench: jobs worked [     112316 ], inserted [          0 ], job/sec [    56158.0 ] [2s]
May 17 14:27:48.977 WRN jobcompleter.BatchCompleter: Hit maximum backlog; completions will wait until below threshold max_backlog=20000
bench: jobs worked [     102054 ], inserted [          0 ], job/sec [    51027.0 ] [2s]
May 17 14:27:50.469 WRN jobcompleter.BatchCompleter: Hit maximum backlog; completions will wait until below threshold max_backlog=20000
May 17 14:27:51.667 WRN jobcompleter.BatchCompleter: Hit maximum backlog; completions will wait until below threshold max_backlog=20000
bench: jobs worked [     111327 ], inserted [          0 ], job/sec [    55663.5 ] [2s]
May 17 14:27:52.696 WRN jobcompleter.BatchCompleter: Hit maximum backlog; completions will wait until below threshold max_backlog=20000
May 17 14:27:54.175 WRN jobcompleter.BatchCompleter: Hit maximum backlog; completions will wait until below threshold max_backlog=20000
bench: jobs worked [     114242 ], inserted [          0 ], job/sec [    57121.0 ] [2s]
May 17 14:27:55.626 WRN jobcompleter.BatchCompleter: Hit maximum backlog; completions will wait until below threshold max_backlog=20000
bench: jobs worked [     122061 ], inserted [          0 ], job/sec [    61030.5 ] [2s]
May 17 14:27:56.911 WRN jobcompleter.BatchCompleter: Hit maximum backlog; completions will wait until below threshold max_backlog=20000
bench: jobs worked [     108867 ], inserted [          0 ], job/sec [    54433.5 ] [2s]
May 17 14:27:58.618 WRN jobcompleter.BatchCompleter: Hit maximum backlog; completions will wait until below threshold max_backlog=20000
bench: jobs worked [     110399 ], inserted [          0 ], job/sec [    55199.5 ] [2s]
bench: total jobs worked [    1000000 ], total jobs inserted [    1000000 ], overall job/sec [    55751.7 ], running 17.936661625s
```

### With change:

```
➜  river git:(bg-optimize-completion-query) ✗ go run ./cmd/river bench --database-url "postgres://localhost/river_test" --num-total-jobs 1_000_000
bench: jobs worked [          0 ], inserted [    1000000 ], job/sec [        0.0 ] [0s]
May 17 14:29:58.613 WRN jobcompleter.BatchCompleter: Hit maximum backlog; completions will wait until below threshold max_backlog=20000
May 17 14:29:59.806 WRN jobcompleter.BatchCompleter: Hit maximum backlog; completions will wait until below threshold max_backlog=20000
bench: jobs worked [     125322 ], inserted [          0 ], job/sec [    62661.0 ] [2s]
May 17 14:30:00.894 WRN jobcompleter.BatchCompleter: Hit maximum backlog; completions will wait until below threshold max_backlog=20000
bench: jobs worked [     128115 ], inserted [          0 ], job/sec [    64057.5 ] [2s]
May 17 14:30:02.264 WRN jobcompleter.BatchCompleter: Hit maximum backlog; completions will wait until below threshold max_backlog=20000
May 17 14:30:03.531 WRN jobcompleter.BatchCompleter: Hit maximum backlog; completions will wait until below threshold max_backlog=20000
bench: jobs worked [     134544 ], inserted [          0 ], job/sec [    67272.0 ] [2s]
May 17 14:30:04.930 WRN jobcompleter.BatchCompleter: Hit maximum backlog; completions will wait until below threshold max_backlog=20000
bench: jobs worked [     128138 ], inserted [          0 ], job/sec [    64069.0 ] [2s]
May 17 14:30:06.822 WRN jobcompleter.BatchCompleter: Hit maximum backlog; completions will wait until below threshold max_backlog=20000
bench: jobs worked [     139629 ], inserted [          0 ], job/sec [    69814.5 ] [2s]
May 17 14:30:09.124 WRN jobcompleter.BatchCompleter: Hit maximum backlog; completions will wait until below threshold max_backlog=20000
bench: jobs worked [     132242 ], inserted [          0 ], job/sec [    66121.0 ] [2s]
bench: jobs worked [     138071 ], inserted [          0 ], job/sec [    69035.5 ] [2s]
May 17 14:30:12.062 WRN jobcompleter.BatchCompleter: Hit maximum backlog; completions will wait until below threshold max_backlog=20000
bench: jobs worked [      73939 ], inserted [          0 ], job/sec [    36969.5 ] [2s]
bench: total jobs worked [    1000000 ], total jobs inserted [    1000000 ], overall job/sec [    66571.6 ], running 15.021423208s
```