@bgentry 2025-05-18
@brandur down to one failing sqlite test that seems to be an issue with timezone conversions. You might have a better idea of what's going on there?