@brandur 2025-05-18
> @brandur down to one failing sqlite test that seems to be an issue with timezone conversions. You might have a better idea of what's going on there?

It looks like you got it right?

Usually this stems from use of `time.Now()` in a driver test instead of `time.Now().UTC()`. All the row -> rivertype helpers in drivers look like this using `UTC()` all over the place:

``` go
	return &rivertype.JobRow{
		ID:           internal.ID,
		Attempt:      max(int(internal.Attempt), 0),
		AttemptedAt:  attemptedAt,
		AttemptedBy:  internal.AttemptedBy,
		CreatedAt:    internal.CreatedAt.UTC(),
		EncodedArgs:  internal.Args,
		Errors:       errors,
		FinalizedAt:  finalizedAt,
		Kind:         internal.Kind,
		MaxAttempts:  max(int(internal.MaxAttempts), 0),
		Metadata:     internal.Metadata,
		Priority:     max(int(internal.Priority), 0),
		Queue:        internal.Queue,
		ScheduledAt:  internal.ScheduledAt.UTC(),
		State:        rivertype.JobState(internal.State),
		Tags:         internal.Tags,
		UniqueKey:    internal.UniqueKey,
		UniqueStates: uniquestates.UniqueBitmaskToStates(uniqueStatesByte),
	}, nil
```

Postgres has timezones in its timestamps, so it can tolerate a time being put in on one timezone, keep track of timezone, and then compare to another time in a different zone, but SQLite is much less tolerant of all of that.