@bgentry 2025-05-18
> It looks like you got it right?

No, actually, it just passes in CI bc it's on UTC :) I fixed it now locally and in CI.

> Usually this stems from use of `time.Now()` in a driver test instead of `time.Now().UTC()`. All the row -> rivertype helpers in drivers look like this using `UTC()` all over the place:

I found a handful of places in the driver that were sending timestamps to the DB _without_ converting them to UTC, including the helpers `timeString` and `timeStringNullable`. I updated these to consistently use UTC as that appears to be what's required, though I'm wondering if there might be some missing test coverage on some of these or reasons certain ones _don't_ need to be converted to UTC first.