@bgentry 2025-05-18
@brandur I don't think there's any risk to merging here w/ the sqlite changes I made but you should definitely check them out and see if I did something dumb or unnecessary, or if we're missing test coverage and I fixed real issues. I for sure had to make the change for `JobInsertFullMany` to get the insert to use the right timestamp on `attempted_at`, and then another couple tests broke in response to the `timeString` change which I fixed with `.UTC()` additions. And then a few I added just after skimming the driver file despite no breakage on either side of the change.